pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag10kq"
version = "0.1.0"
description = "RAG system for 10-K and 10-Q filings"
requires-python = ">=3.8"
dependencies = [
    "beautifulsoup4==4.12.3",
    "lxml==5.3.0",
    "pandas==2.2.2",
    "markdownify==0.13.1",
    "numpy==1.26.4",
    "scikit-learn==1.5.2",
    "tqdm==4.66.5",
    "faiss-cpu==1.8.0.post1",
    "rank-bm25==0.2.2",
    "sentence-transformers==3.1.1",
]

[tool.setuptools.packages.find]
where = ["src"]
